requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.11.16
aiosignal==1.3.2
annotated-types==0.7.0
anyio==4.9.0
appdirs==1.4.4
attrs==25.3.0
certifi==2025.1.31
charset-normalizer==3.4.1
dataclasses-json==0.6.7
datasets==3.5.0
dill==0.3.8
diskcache==5.6.3
distro==1.9.0
elastic-transport==8.17.1
elasticsearch==8.17.2
filelock==3.18.0
frozenlist==1.5.0
fsspec==2024.12.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
httpx-sse==0.4.0
huggingface-hub==0.30.1
idna==3.10
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonpatch==1.33
jsonpointer==3.0.0
langchain==0.3.23
langchain-community==0.3.21
langchain-core==0.3.51
langchain-elasticsearch==0.3.2
langchain-huggingface==0.1.2
langchain-openai==0.3.12
langchain-text-splitters==0.3.8
langsmith==0.3.24
MarkupSafe==3.0.2
marshmallow==3.26.1
mpmath==1.3.0
multidict==6.3.2
multiprocess==0.70.16
mypy-extensions==1.0.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.4
ollama==0.4.7
openai==1.70.0
orjson==3.10.16
packaging==24.2
pandas==2.2.3
pillow==11.1.0
propcache==0.3.1
pyarrow==19.0.1
pydantic==2.11.2
pydantic-settings==2.8.1
pydantic_core==2.33.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.0
pytz==2025.2
PyYAML==6.0.2
ragas==0.2.14
regex==2024.11.6
requests==2.32.3
requests-toolbelt==1.0.0
safetensors==0.5.3
scikit-learn==1.6.1
scipy==1.15.2
sentence-transformers==4.0.2
simsimd==6.2.1
six==1.17.0
sniffio==1.3.1
SQLAlchemy==2.0.40
sympy==1.13.1
tenacity==9.1.2
threadpoolctl==3.6.0
tiktoken==0.9.0
tokenizers==0.21.1
torch==2.6.0
tqdm==4.67.1
transformers==4.50.3
typing-inspect==0.9.0
typing-inspection==0.4.0
typing_extensions==4.13.1
tzdata==2025.2
urllib3==2.3.0
xxhash==3.5.0
yarl==1.18.3
zstandard==0.23.0
